Update profile to be compatible with CoRIM changes
The instance-id-type-choice is not allowed to have its own comparison
algorithm, so drop the different "notions of identity" in favor of
supporting just VCEK- and VLEK-based identity.

This means the REPORT_ID and REPORT_ID_MA are added to the
measurement-map structure and corresponding translation of evidence.

The ID_KEY_DIGEST and AUTHOR_KEY_DIGEST are added to measurements to
allow for their use directly in conditional endorsements. I've made
their interpretation as reference value optional, since I think we want
to encourage reference value providers to issue CoRIMs, but it can be
used for simpler developments.