Separate mkey and element-id text
The flag interpretation in the evidence transformation was too cramped,
so it is moved to the section about the extensions.

The `mkey` descriptions for CoRIMs and the `element-id` descriptions for
the evidence transformations are put in their own sections. The use of
mkey is used for expressing reference values and endorsements, whereas
the element-id for the transformation is meant for ingesting evidence
that is compatible with reference values and endorsements.